NMS-9128: Simplify AssetGraphMLProvider
